Playing with ssl options
git-svn-id: http://dcplusplus.svn.sourceforge.net/svnroot/dcplusplus/dcplusplus/trunk@624 f586c4b0-9d0d-0410-bc51-f1fe239209fc

+++ client/BufferedSocket.cpp
@@ -255,7 +255,7 @@ void BufferedSocket::threadRead() throw(SocketException) {
 		}
 	}
 
-	if(mode == MODE_LINE && line.size() > SETTING(MAX_COMMAND_LENGTH)) {
+	if(mode == MODE_LINE && line.size() > static_cast<size_t>(SETTING(MAX_COMMAND_LENGTH))) {
 		throw SocketException(STRING(COMMAND_TOO_LONG));
 	}
 }
